Rename ObjectKeyInterface to ObjectIdentityInterface.
