+++ views/extension/_sidebar.php
@@ -1,6 +1,6 @@
 <?php
 
-use app\components\objectKey\ClassType;
+use app\components\object\ClassType;
 use app\models\ExtensionCategory;
 use app\models\ExtensionTag;
 use app\widgets\RecentComments;
